[charts] Refactor base for api and schemas
