+++ superset/views/base.py
@@ -18,13 +18,12 @@ import functools
 import logging
 import traceback
 from datetime import datetime
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, Optional
 
 import simplejson as json
 import yaml
 from flask import (
     abort,
-    current_app,
     flash,
     g,
     get_flashed_messages,
@@ -33,19 +32,14 @@ from flask import (
     Response,
     session,
 )
-from flask_appbuilder import BaseView, Model, ModelRestApi, ModelView
+from flask_appbuilder import BaseView, ModelView
 from flask_appbuilder.actions import action
-from flask_appbuilder.api import expose, protect, rison, safe
 from flask_appbuilder.forms import DynamicForm
-from flask_appbuilder.models.filters import Filters
 from flask_appbuilder.models.sqla.filters import BaseFilter
 from flask_appbuilder.widgets import ListWidget
 from flask_babel import get_locale, gettext as __, lazy_gettext as _
 from flask_wtf.form import FlaskForm
-from marshmallow import post_load, pre_load, Schema, ValidationError
 from sqlalchemy import or_
-from sqlalchemy.exc import SQLAlchemyError
-from sqlalchemy.orm.exc import NoResultFound
 from werkzeug.exceptions import HTTPException
 from wtforms.fields.core import Field, UnboundField
 
@@ -163,26 +157,6 @@ def handle_api_exception(f):
             return json_error_response(
                 utils.error_msg_from_exception(e), stacktrace=utils.get_stacktrace()
             )
-
-    return functools.update_wrapper(wraps, f)
-
-
-def check_ownership_and_item_exists(f):
-    """
-    A Decorator that checks if an object exists and is owned by the current user
-    """
-
-    def wraps(self, pk):  # pylint: disable=invalid-name
-        item = self.datamodel.get(
-            pk, self._base_filters  # pylint: disable=protected-access
-        )
-        if not item:
-            return self.response_404()
-        try:
-            check_ownership(item)
-        except SupersetSecurityException as e:
-            return self.response(403, message=str(e))
-        return f(self, item)
 
     return functools.update_wrapper(wraps, f)
 
@@ -388,364 +362,6 @@ class DatasourceFilter(BaseFilter):  # pylint: disable=too-few-public-methods
                 self.model.schema_perm.in_(schema_perms),
             )
         )
-
-
-def validate_owner(value):
-    try:
-        (
-            current_app.appbuilder.get_session.query(
-                current_app.appbuilder.sm.user_model.id
-            )
-            .filter_by(id=value)
-            .one()
-        )
-    except NoResultFound:
-        raise ValidationError(f"User {value} does not exist")
-
-
-class BaseSupersetSchema(Schema):
-    """
-    Extends Marshmallow schema so that we can pass a Model to load
-    (following marshamallow-sqlalchemy pattern). This is useful
-    to perform partial model merges on HTTP PUT
-    """
-
-    __class_model__: Model = None
-
-    def __init__(self, **kwargs):
-        self.instance: Optional[Model] = None
-        super().__init__(**kwargs)
-
-    def load(
-        self, data, many=None, partial=None, instance: Model = None, **kwargs
-    ):  # pylint: disable=arguments-differ
-        self.instance = instance
-        return super().load(data, many=many, partial=partial, **kwargs)
-
-    @post_load
-    def make_object(self, data: Dict, discard: List[str] = None) -> Model:
-        """
-        Creates a Model object from POST or PUT requests. PUT will use self.instance
-        previously fetched from the endpoint handler
-
-        :param data: Schema data payload
-        :param discard: List of fields to not set on the model
-        """
-        discard = discard or []
-        if not self.instance:
-            self.instance = self.__class_model__()  # pylint: disable=not-callable
-        for field in data:
-            if field not in discard:
-                setattr(self.instance, field, data.get(field))
-        return self.instance
-
-
-class BaseOwnedSchema(BaseSupersetSchema):
-    """
-    Implements owners validation and pre load on Marshmallow schemas
-    """
-
-    owners_field_name = "owners"
-
-    @post_load
-    def make_object(self, data: Dict, discard: List[str] = None) -> Model:
-        discard = discard or []
-        discard.append(self.owners_field_name)
-        instance = super().make_object(data, discard)
-        if "owners" not in data and g.user not in instance.owners:
-            instance.owners.append(g.user)
-        if self.owners_field_name in data:
-            self.set_owners(instance, data[self.owners_field_name])
-        return instance
-
-    @pre_load
-    def pre_load(self, data: Dict):
-        data[self.owners_field_name] = data.get(self.owners_field_name, [])
-
-    @staticmethod
-    def set_owners(instance: Model, owners: List[int]):
-        owner_objs = list()
-        if g.user.id not in owners:
-            owners.append(g.user.id)
-        for owner_id in owners:
-            user = current_app.appbuilder.get_session.query(
-                current_app.appbuilder.sm.user_model
-            ).get(owner_id)
-            owner_objs.append(user)
-        instance.owners = owner_objs
-
-
-get_related_schema = {
-    "type": "object",
-    "properties": {
-        "page_size": {"type": "integer"},
-        "page": {"type": "integer"},
-        "filter": {"type": "string"},
-    },
-}
-
-
-class BaseSupersetModelRestApi(ModelRestApi):
-    """
-    Extends FAB's ModelResApi to implement specific superset generic functionality
-    """
-
-    order_rel_fields: Dict[str, Tuple[str, str]] = {}
-    """
-    Impose ordering on related fields query::
-
-        order_rel_fields = {
-            "<RELATED_FIELD>": ("<RELATED_FIELD_FIELD>", "<asc|desc>"),
-             ...
-        }
-    """  # pylint: disable=pointless-string-statement
-    filter_rel_fields_field: Dict[str, str] = {}
-    """
-    Declare the related field field for filtering::
-
-        filter_rel_fields_field = {
-            "<RELATED_FIELD>": "<RELATED_FIELD_FIELD>", "<asc|desc>")
-        }
-    """  # pylint: disable=pointless-string-statement
-
-    def _get_related_filter(self, datamodel, column_name: str, value: str) -> Filters:
-        filter_field = self.filter_rel_fields_field.get(column_name)
-        filters = datamodel.get_filters([filter_field])
-        if value:
-            filters.rest_add_filters(
-                [{"opr": "sw", "col": filter_field, "value": value}]
-            )
-        return filters
-
-    @expose("/related/<column_name>", methods=["GET"])
-    @protect()
-    @safe
-    @rison(get_related_schema)
-    def related(self, column_name: str, **kwargs):
-        """Get related fields data
-        ---
-        get:
-          parameters:
-          - in: path
-            schema:
-              type: string
-            name: column_name
-          - in: query
-            name: q
-            content:
-              application/json:
-                schema:
-                  type: object
-                  properties:
-                    page_size:
-                      type: integer
-                    page:
-                      type: integer
-                    filter:
-                      type: string
-          responses:
-            200:
-              description: Related column data
-              content:
-                application/json:
-                  schema:
-                    type: object
-                    properties:
-                      count:
-                        type: integer
-                      result:
-                        type: object
-                        properties:
-                          value:
-                            type: integer
-                          text:
-                            type: string
-            400:
-              $ref: '#/components/responses/400'
-            401:
-              $ref: '#/components/responses/401'
-            404:
-              $ref: '#/components/responses/404'
-            422:
-              $ref: '#/components/responses/422'
-            500:
-              $ref: '#/components/responses/500'
-        """
-        args = kwargs.get("rison", {})
-        # handle pagination
-        page, page_size = self._handle_page_args(args)
-        try:
-            datamodel = self.datamodel.get_related_interface(column_name)
-        except KeyError:
-            return self.response_404()
-        page, page_size = self._sanitize_page_args(page, page_size)
-        # handle ordering
-        order_field = self.order_rel_fields.get(column_name)
-        if order_field:
-            order_column, order_direction = order_field
-        else:
-            order_column, order_direction = "", ""
-        # handle filters
-        filters = self._get_related_filter(datamodel, column_name, args.get("filter"))
-        # Make the query
-        count, values = datamodel.query(
-            filters, order_column, order_direction, page=page, page_size=page_size
-        )
-        # produce response
-        result = [
-            {"value": datamodel.get_pk_value(value), "text": str(value)}
-            for value in values
-        ]
-        return self.response(200, count=count, result=result)
-
-
-class BaseOwnedModelRestApi(BaseSupersetModelRestApi):
-    @expose("/<pk>", methods=["PUT"])
-    @protect()
-    @check_ownership_and_item_exists
-    @safe
-    def put(self, item):  # pylint: disable=arguments-differ
-        """Changes a owned Model
-        ---
-        put:
-          parameters:
-          - in: path
-            schema:
-              type: integer
-            name: pk
-          requestBody:
-            description: Model schema
-            required: true
-            content:
-              application/json:
-                schema:
-                  $ref: '#/components/schemas/{{self.__class__.__name__}}.put'
-          responses:
-            200:
-              description: Item changed
-              content:
-                application/json:
-                  schema:
-                    type: object
-                    properties:
-                      result:
-                        $ref: '#/components/schemas/{{self.__class__.__name__}}.put'
-            400:
-              $ref: '#/components/responses/400'
-            401:
-              $ref: '#/components/responses/401'
-            403:
-              $ref: '#/components/responses/401'
-            404:
-              $ref: '#/components/responses/404'
-            422:
-              $ref: '#/components/responses/422'
-            500:
-              $ref: '#/components/responses/500'
-        """
-        if not request.is_json:
-            self.response_400(message="Request is not JSON")
-        item = self.edit_model_schema.load(request.json, instance=item)
-        if item.errors:
-            return self.response_422(message=item.errors)
-        try:
-            self.datamodel.edit(item.data, raise_exception=True)
-            return self.response(
-                200, result=self.edit_model_schema.dump(item.data, many=False).data
-            )
-        except SQLAlchemyError as e:
-            return self.response_422(message=str(e))
-
-    @expose("/", methods=["POST"])
-    @protect()
-    @safe
-    def post(self):
-        """Creates a new owned Model
-        ---
-        post:
-          requestBody:
-            description: Model schema
-            required: true
-            content:
-              application/json:
-                schema:
-                  $ref: '#/components/schemas/{{self.__class__.__name__}}.post'
-          responses:
-            201:
-              description: Model added
-              content:
-                application/json:
-                  schema:
-                    type: object
-                    properties:
-                      id:
-                        type: string
-                      result:
-                        $ref: '#/components/schemas/{{self.__class__.__name__}}.post'
-            400:
-              $ref: '#/components/responses/400'
-            401:
-              $ref: '#/components/responses/401'
-            422:
-              $ref: '#/components/responses/422'
-            500:
-              $ref: '#/components/responses/500'
-        """
-        if not request.is_json:
-            return self.response_400(message="Request is not JSON")
-        item = self.add_model_schema.load(request.json)
-        # This validates custom Schema with custom validations
-        if item.errors:
-            return self.response_422(message=item.errors)
-        try:
-            self.datamodel.add(item.data, raise_exception=True)
-            return self.response(
-                201,
-                result=self.add_model_schema.dump(item.data, many=False).data,
-                id=item.data.id,
-            )
-        except SQLAlchemyError as e:
-            return self.response_422(message=str(e))
-
-    @expose("/<pk>", methods=["DELETE"])
-    @protect()
-    @check_ownership_and_item_exists
-    @safe
-    def delete(self, item):  # pylint: disable=arguments-differ
-        """Deletes owned Model
-        ---
-        delete:
-          parameters:
-          - in: path
-            schema:
-              type: integer
-            name: pk
-          responses:
-            200:
-              description: Model delete
-              content:
-                application/json:
-                  schema:
-                    type: object
-                    properties:
-                      message:
-                        type: string
-            401:
-              $ref: '#/components/responses/401'
-            403:
-              $ref: '#/components/responses/401'
-            404:
-              $ref: '#/components/responses/404'
-            422:
-              $ref: '#/components/responses/422'
-            500:
-              $ref: '#/components/responses/500'
-        """
-        try:
-            self.datamodel.delete(item, raise_exception=True)
-            return self.response(200, message="OK")
-        except SQLAlchemyError as e:
-            return self.response_422(message=str(e))
 
 
 class CsvResponse(Response):  # pylint: disable=too-many-ancestors
