+++ superset/views/chart/api.py
@@ -28,7 +28,8 @@ from superset.exceptions import SupersetException
 from superset.models.dashboard import Dashboard
 from superset.models.slice import Slice
 from superset.utils import core as utils
-from superset.views.base import BaseOwnedModelRestApi, BaseOwnedSchema, validate_owner
+from superset.views.base_api import BaseOwnedModelRestApi
+from superset.views.base_schemas import BaseOwnedSchema, validate_owner
 from superset.views.chart.mixin import SliceMixin
