+++ superset/views/dashboard/api.py
@@ -27,7 +27,8 @@ from superset import appbuilder
 from superset.exceptions import SupersetException
 from superset.models.dashboard import Dashboard
 from superset.utils import core as utils
-from superset.views.base import BaseOwnedModelRestApi, BaseOwnedSchema, validate_owner
+from superset.views.base_api import BaseOwnedModelRestApi
+from superset.views.base_schemas import BaseOwnedSchema, validate_owner
 
 from .mixin import DashboardMixin
